Fixes minor mistakes.
