+++ src/Compiler/SetCompiler.php
@@ -27,7 +27,6 @@ namespace Datto\Cinnabari\Compiler;
 use Datto\Cinnabari\Exception\ArgumentsException;
 use Datto\Cinnabari\Exception\CompilerException;
 use Datto\Cinnabari\Format\Arguments;
-use Datto\Cinnabari\Mysql\Expression\AbstractExpression;
 use Datto\Cinnabari\Mysql\Expression\Column;
 use Datto\Cinnabari\Mysql\Expression\Parameter;
 use Datto\Cinnabari\Mysql\Update;
@@ -146,6 +145,41 @@ class SetCompiler extends AbstractCompiler
 
             $this->mysql->addPropertyValuePair($this->context, $column, $expression);
         }
+
+        return true;
+    }
+
+    protected function getOptionalSortFunction()
+    {
+        if (!self::scanFunction(reset($this->request), $name, $arguments)) {
+            return false;
+        }
+
+        if ($name !== 'sort') {
+            return false;
+        }
+
+        // at this point, we're sure they want to sort
+        if (!isset($arguments) || count($arguments) !== 1) {
+            // TODO: add an explanation of the missing argument, or link to the documentation
+            throw CompilerException::noSortArguments($this->request);
+        }
+
+        $state = array($this->request, $this->context);
+
+        // consume all of the joins
+        $this->request = $arguments[0];
+        $this->request = $this->followJoins($this->request);
+
+        if (!$this->scanProperty(reset($this->request), $table, $name, $type, $hasZero)) {
+            return false;
+        }
+
+        $this->mysql->setOrderBy($this->context, $name, true);
+
+        list($this->request, $this->context) = $state;
+
+        array_shift($this->request);
 
         return true;
     }
